+++ Il2CppAssemblyGenerator/Packages/Cpp2IL.cs
@@ -8,7 +8,7 @@ namespace MelonLoader.Il2CppAssemblyGenerator
         {
             Version = MelonCommandLine.AssemblyGenerator.ForceVersion_Dumper;
             if (string.IsNullOrEmpty(Version) || Version.Equals("0.0.0.0"))
-                Version = RemoteAPI.LAST_RESPONSE.ForceDumperVersion;
+                Version = RemoteAPI.ReturnedInfo.ForceDumperVersion;
             if (string.IsNullOrEmpty(Version) || Version.Equals("0.0.0.0"))
                 Version = "commit_1aba7eee0667903dd465d897d88df15c2fe8b991";
             URL = "https://github.com/SamboyCoding/Cpp2IL/releases/download/" + Version + "/Cpp2IL-Win.exe";