+++ MelonLoader/Extensions/Il2CppAssemblyGenerator.cs
@@ -6,55 +6,66 @@ namespace MelonLoader
 {
     internal static class Il2CppAssemblyGenerator
     {
+        private static string FileNameWithExtension = "Il2CppAssemblyGenerator.dll";
+        private static Assembly asm = null;
         private static MethodInfo RunMethod = null;
+        internal static Assembly AssemblyResolver(object sender, ResolveEventArgs args) => args.Name.StartsWith($"{Path.GetFileNameWithoutExtension(FileNameWithExtension)}, Version=") ? asm : null;
 
         internal static bool Run()
         {
-            MelonLogger.Msg("Loading Il2CppAssemblyGenerator...");
-            if (RunMethod == null)
-            {
-                string BaseDirectory = Path.Combine(Path.Combine(Path.Combine(MelonUtils.GameDirectory, "MelonLoader"), "Dependencies"), "Il2CppAssemblyGenerator");
-                if (!Directory.Exists(BaseDirectory))
-                {
-                    MelonLogger.Error("Failed to Find Il2CppAssemblyGenerator Directory!");
-                    return false;
-                }
-                string AssemblyPath = Path.Combine(BaseDirectory, "Il2CppAssemblyGenerator.dll");
-                if (!File.Exists(AssemblyPath))
-                {
-                    MelonLogger.Error("Failed to Find Il2CppAssemblyGenerator.dll!");
-                    return false;
-                }
-                try
-                {
-                    Assembly asm = Assembly.LoadFrom(AssemblyPath);
-                    if (asm == null)
-                    {
-                        MelonLogger.ThrowInternalFailure($"Failed to Load Assembly for Il2CppAssemblyGenerator.dll!");
-                        return false;
-                    }
-                    Type type = asm.GetType("MelonLoader.Il2CppAssemblyGenerator.Core");
-                    if (type == null)
-                    {
-                        MelonLogger.ThrowInternalFailure($"Failed to Get Type for MelonLoader.Il2CppAssemblyGenerator.Core!");
-                        return false;
-                    }
-                    RunMethod = type.GetMethod("Run", BindingFlags.NonPublic | BindingFlags.Static);
-                }
-                catch (Exception ex)
-                {
-                    MelonLogger.ThrowInternalFailure($"Il2CppAssemblyGenerator Exception: {ex}");
-                    return false;
-                }
-            }
+            if (!MelonUtils.IsGameIl2Cpp())
+                return true;
+
             if (RunMethod != null)
             {
                 int returnval = (int)RunMethod.Invoke(null, new object[0]);
                 Fixes.ApplicationBase.Run(AppDomain.CurrentDomain);
-                return returnval == 0;
+                return (returnval == 0);
             }
-            MelonLogger.ThrowInternalFailure($"Failed to Get Run Method for MelonLoader.AssemblyGenerator.Core!");
+
             return false;
+        }
+
+        internal static void Load()
+        {
+            if (!MelonUtils.IsGameIl2Cpp())
+                return;
+
+            MelonLogger.Msg("Loading Il2CppAssemblyGenerator...");
+
+            string BaseDirectory = Path.Combine(Path.Combine(Path.Combine(MelonUtils.GameDirectory, "MelonLoader"), "Dependencies"), "Il2CppAssemblyGenerator");
+            if (!Directory.Exists(BaseDirectory))
+            {
+                MelonLogger.Error("Failed to Find Il2CppAssemblyGenerator Directory!");
+                return;
+            }
+
+            string AssemblyPath = Path.Combine(BaseDirectory, FileNameWithExtension);
+            if (!File.Exists(AssemblyPath))
+            {
+                MelonLogger.Error($"Failed to Find {FileNameWithExtension}!");
+                return;
+            }
+
+            try
+            {
+                asm = Assembly.LoadFrom(AssemblyPath);
+                if (asm == null)
+                {
+                    MelonLogger.ThrowInternalFailure($"Failed to Load Assembly for {FileNameWithExtension}!");
+                    return;
+                }
+
+                Type type = asm.GetType("MelonLoader.Il2CppAssemblyGenerator.Core");
+                if (type == null)
+                {
+                    MelonLogger.ThrowInternalFailure($"Failed to Get Type for MelonLoader.Il2CppAssemblyGenerator.Core!");
+                    return;
+                }
+
+                RunMethod = type.GetMethod("Run", BindingFlags.NonPublic | BindingFlags.Static);
+            }
+            catch (Exception ex) { MelonLogger.ThrowInternalFailure($"Il2CppAssemblyGenerator Exception: {ex}"); }
         }
     }
 }