+++ Proxy/Base/DllMain.cpp
@@ -4,6 +4,7 @@ BOOL WINAPI DllMain(HINSTANCE hinstDLL, DWORD fdwReason, LPVOID lpvReserved)
 {
 	if (fdwReason != DLL_PROCESS_ATTACH)
 		return TRUE;
+	DisableThreadLibraryCalls(hinstDLL);
 	Core::Load(hinstDLL);
 	return TRUE;
 }